requirements.txt
langchain
python-dotenv

langchain_community
# pypdf
# bs4
# arxiv
# PyMuPDF
# wikipedia
langchain-text-splitters
langchain-community
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_ollama
# fastapi
# uvicorn
langserve
langchain-groq